% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCgampoi.R
\name{gp.lr}
\alias{gp.lr}
\title{Structure learning with negative binomial model using the glmGamPoi package}
\usage{
gp.lr(X, maxcard, alpha, extend, overdispersion)
}
\arguments{
\item{X}{the matrix of counts (n times p).}

\item{maxcard}{the uper bound of the cardinality of the conditional sets K}

\item{alpha}{the significant level of the tests}

\item{extend}{if TRUE it considers the union of the tests, otherwise it
considers the intersection.}

\item{overdispersion}{if TRUE a negative binomial is fit, if FALSE a Poisson.}
}
\value{
the estimated adjacency matrix of the graph.
}
\description{
This function estimates the adjacency matrix of a NB model given a matrix of
counts, using the glm_gp function.
}
