% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPGMguide.R
\name{LPGM.ord}
\alias{LPGM.ord}
\title{Structure learning with Poisson models using the Or-LPGM algorithm}
\usage{
LPGM.ord(X, order, alpha)
}
\arguments{
\item{X}{the matrix of counts (n times p).}

\item{order}{the topological ordering of variables (names of nodes).}

\item{alpha}{the significant level of the tests.}
}
\value{
the estimated adjacency matrix of the graph.
}
\description{
This function estimates the adjacency matrix of a Local Poisson Graphical
Model (LPGM) given a matrix of counts and topological ordering, using
generalized linear models. It is a natural extension of the LPGM of Allen and
Liu (2013). See Nguyen et al. (2022) for details.
}
\references{
Allen and Liu (2013). A local Poisson graphical model for inferring
networksfrom sequencing data. IEEE Transactions on NanoBioscience, 12(3),
189–198.

Nguyen, Chiogna, Risso, Banzato (2022). Guided structure learning of
DAGs for count data. arXiv:2206.09754.
}
