% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasimNB.R
\name{nbinom.Simdata}
\alias{nbinom.Simdata}
\title{Generate negatibe binomial (NB) data}
\usage{
nbinom.Simdata(n, p, B, mu, mu.nois, theta)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of variables (nodes)}

\item{B}{adjacency matrix}

\item{mu}{mean of negative binomial}

\item{mu.nois}{mean of noise}

\item{theta}{dispersion parameter of negative binomial, where \eqn{var=\mu+\mu^2/theta}}
}
\description{
Generatedata matrix (nxp) under NB assumption based on Karlis (or LPGM. 2013) with adjancency matrix B.
}
