% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCguide.R
\name{pois.ord}
\alias{pois.ord}
\title{Structure learning with Poisson models using the Or-PPGM algorithm}
\usage{
pois.ord(X, maxcard, alpha, order)
}
\arguments{
\item{X}{the matrix of counts (n times p).}

\item{maxcard}{the uper bound of the cardinality of the conditional sets K.}

\item{alpha}{the significant level of the tests.}

\item{order}{the topological ordering of variables (names of nodes).}
}
\value{
the estimated adjacency matrix of the graph.
}
\description{
This function estimates the adjacency matrix of a Poisson model given a
matrix of counts and topological ordering, using the Or-PPGM algorithm
proposed in Nguyen et al. (2022).
}
\references{
Nguyen, Chiogna, Risso, Banzato (2022). Guided structure learning of
DAGs for count data. arXiv:2206.09754.
}
