% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasimPois.R
\name{pois.simdata}
\alias{pois.simdata}
\title{Generate Poisson data}
\usage{
pois.simdata(n, p, B, lambda, lambda.c)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of variables (nodes)}

\item{B}{adjacency matrix}

\item{lambda}{mean of Poisson}

\item{lambda.c}{mean of noise}
}
\description{
Generate data matrix (nxp) under Poisson assumption based on Karlis (or LPGM. 2013) with adjancency matrix B,
}
