% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCPoisson.R
\name{pois.wald}
\alias{pois.wald}
\title{This function estimates adjacency matrix of a Poisson models given a matrix of counts, using glm.}
\usage{
pois.wald(X, maxcard, alpha, extend)
}
\arguments{
\item{X}{the matrix of counts.}

\item{maxcard}{the uper bound for cardinalities of conditional sets K
extend=TRUE if we consider the union of tests}

\item{alpha}{the sisnificant level tests}
}
\value{
the adj matrix.
}
\description{
This function estimates adjacency matrix of a Poisson models given a matrix of counts, using glm.
}
