% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{prediction_scores}
\alias{prediction_scores}
\title{Prediction scores for estimated graph}
\usage{
prediction_scores(trueG, estimatedG)
}
\arguments{
\item{trueG}{the adjacency matrix of the true graph.}

\item{estimatedG}{the adjacency matrix of the estimated graph.}
}
\value{
A vector with the number of true positives (TP), the number of false
  positives (FP), the number of false negatives (FN), the positive predictive
  value (PPV), the sensitivity (Se), and the F1 score (F1).
}
\description{
This function returns a set of metrics useful to evaluate the goodness of the
estimation of the graph on simulated data, or whenever a true graph is
available.
}
\examples{
adj <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0), nrow=3)
mat <- simdata(n=10, p=3, B=adj, mu=5, mu_noise=1)
res <- PCzinb(mat, method="poi")
prediction_scores(adj, res)
}
