% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasim.R
\name{simdata}
\alias{simdata}
\title{Generate Poisson, negative binomial (NB), and zero-inflated NB (ZINB) data}
\usage{
simdata(
  n,
  p,
  B,
  family = c("Poisson", "NB", "ZINB"),
  mu,
  mu_noise,
  theta = NA,
  pi = NA
)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of variables (nodes)}

\item{B}{adjacency matrix}

\item{family}{the distribution from which to simulate (Poisson, NB, ZINB)}

\item{mu}{mean of the non-zero component}

\item{mu_noise}{mean of noise}

\item{theta}{dispersion parameter of the non-zero component,
where \eqn{var=\mu+\mu^2/theta}}

\item{pi}{probability of zero inflation}
}
\description{
This function generates a data matrix (\eqn{n \times p}) under Poisson, NB,
or ZINB assumptions, based on Karlis (or LPGM. 2013) with adjancency matrix B.
}
\examples{
p <- 4
B <- matrix(c(0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 1, 0, 0), nrow=p)
n <- 10
simdata(n, p, B, mu = 5, mu_noise=1)
simdata(n, p, B, mu = 5, mu_noise=1, family="NB", theta=1)
simdata(n, p, B, mu = 5, mu_noise=1, family="ZINB", theta=1, pi=0.2)
}
