% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasimzinb.R
\name{zinb.simdata}
\alias{zinb.simdata}
\title{Generate zero-inflated negative binomial data}
\usage{
zinb.simdata(n, p, B, mu, mu.nois, theta, pi)
}
\arguments{
\item{n}{number of samples}

\item{p}{number of variables (nodes)}

\item{B}{adjacency matrix}

\item{mu}{mean of negative binomial part}

\item{mu.nois}{mean of noise}

\item{theta}{dispersion parameter of negative binomial part, where \eqn{var=\mu+\mu^2/theta}}

\item{pi}{parameter relate to zero part and negative part}
}
\description{
Simulate zinb data based on Karlis (or LPGM)
}
