% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCzinb1noT.R
\name{zinb1.noT}
\alias{zinb1.noT}
\title{This function estimates adjacency matrix of a zinb1 models given a matrix of counts, using optim.
zinb1 models are zinb models with the structure is infered from both mu and pi}
\usage{
zinb1.noT(X, maxcard, alpha, extend)
}
\arguments{
\item{X}{the matrix of counts.}

\item{maxcard}{the uper bound for cardinalities of conditional sets K
extend=TRUE if we consider the union of tests}

\item{alpha}{the sisnificant level tests}
}
\value{
the adj matrix.
}
\description{
This function estimates adjacency matrix of a zinb1 models given a matrix of counts, using optim.
zinb1 models are zinb models with the structure is infered from both mu and pi
}
