% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCnb_optim.R
\name{nbscale.noT}
\alias{nbscale.noT}
\title{Structure learning with negative binomial model using optim}
\usage{
nbscale.noT(X, maxcard, alpha, extend)
}
\arguments{
\item{X}{the matrix of counts (n times p).}

\item{maxcard}{the uper bound of the cardinality of the conditional sets K}

\item{alpha}{the significant level of the tests}

\item{extend}{if TRUE it considers the union of the tests, otherwise it
considers the intersection.}
}
\value{
the estimated adjacency matrix of the graph.
}
\description{
This function estimates the adjacency matrix of a NB model given a matrix of
counts, using the optim function.
}
