% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimzinb.R
\name{zinbOptimizeDispersion}
\alias{zinbOptimizeDispersion}
\title{(ZINB) model. The ZINB distribution is parametrized by three
parameters: the mean value and the dispersion of the negative binomial
distribution, and the probability of the zero component.}
\usage{
zinbOptimizeDispersion(mu, logitPi, Y, n)
}
\arguments{
\item{mu}{the vector mean  of the negative binomial}

\item{logitPi}{the vector of logit of the probabilities of the zero component}

\item{Y}{the vector of counts}

\item{n}{length of the returned vector}
}
\description{
(ZINB) model. The ZINB distribution is parametrized by three
parameters: the mean value and the dispersion of the negative binomial
distribution, and the probability of the zero component.
}
\examples{
n <- 10
mu <- seq(10,50,length.out=n)
logitPi <- rnorm(1)
zeta <- rnorm(1)
Y <- rnbinom(n=n, size=exp(zeta), mu=mu)
zinbgraph:::zinbOptimizeDispersion ( mu, logitPi,Y,n)
}
